wip moving to adapters
