+++ lib/generators/with_ember_app/initializer_generator.rb
@@ -5,8 +5,6 @@ module WithEmberApp
     class InitializerGenerator < Rails::Generators::Base
       source_root File.expand_path('../../templates', __FILE__)
 
-      puts File.expand_path('../../templates', __FILE__)
-
       desc 'Creates a sample WithEmberApp initializer.'
 
       def create_initializer
